# public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"Brand": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"url": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					},
					"updated": {
						"type": "string",
						"format": "date-time"
					},
					"createdBy": {
						"type": "string"
					},
					"updatedBy": {
						"type": "string"
					},
					"products": {
						"items": {
							"$ref": "#/components/schemas/Product"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProductTypeCategory": {
				"enum": [
					"Bikes",
					"Rider Accessories",
					"Bike Accessories",
					"Components"
				],
				"type": "string"
			},
			"Product": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"lastUsedValue": {
						"type": "number",
						"format": "double"
					},
					"year": {
						"type": "number",
						"format": "double"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					},
					"updated": {
						"type": "string",
						"format": "date-time"
					},
					"createdBy": {
						"type": "string"
					},
					"updatedBy": {
						"type": "string"
					},
					"featured": {
						"type": "boolean"
					},
					"brand": {
						"$ref": "#/components/schemas/Brand"
					},
					"productType": {
						"$ref": "#/components/schemas/ProductType"
					},
					"productDetail": {
						"$ref": "#/components/schemas/ProductDetail"
					},
					"lineItems": {
						"items": {
							"$ref": "#/components/schemas/PurchaseLineItem"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"lastUsedValue",
					"year",
					"featured",
					"brand",
					"productType"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ProductType": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"productTypeCategory": {
						"$ref": "#/components/schemas/ProductTypeCategory"
					},
					"products": {
						"items": {
							"$ref": "#/components/schemas/Product"
						},
						"type": "array"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					},
					"updated": {
						"type": "string",
						"format": "date-time"
					},
					"createdBy": {
						"type": "string"
					},
					"updatedBy": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"productTypeCategory"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Category": {
				"enum": [
					"Kids",
					"Rider Accessories",
					"Bike Accessories",
					"City",
					"Road",
					"E-Bikes"
				],
				"type": "string"
			},
			"SizeMaster": {
				"enum": [
					"Small",
					"Medium",
					"Large",
					"Other"
				],
				"type": "string"
			},
			"SuspensionType": {
				"enum": [
					"Rigid",
					"Suspension",
					"Hydraulic"
				],
				"type": "string"
			},
			"Material": {
				"enum": [
					"Alloy",
					"Steel",
					"Carbon"
				],
				"type": "string"
			},
			"BikeUsers": {
				"enum": [
					"Men",
					"Women",
					"Kids",
					"Unisex"
				],
				"type": "string"
			},
			"ProductDetail": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"gear": {
						"type": "boolean"
					},
					"wheelSize": {
						"type": "number",
						"format": "double"
					},
					"category": {
						"$ref": "#/components/schemas/Category"
					},
					"frameSize": {
						"$ref": "#/components/schemas/SizeMaster"
					},
					"fork": {
						"$ref": "#/components/schemas/SuspensionType"
					},
					"forkMaterial": {
						"$ref": "#/components/schemas/Material"
					},
					"frontBrakes": {
						"type": "string"
					},
					"rearBrakes": {
						"type": "string"
					},
					"productUser": {
						"$ref": "#/components/schemas/BikeUsers"
					},
					"weight": {
						"type": "number",
						"format": "double"
					},
					"crankset": {
						"type": "string"
					},
					"frontDerailer": {
						"type": "string"
					},
					"rearDerailer": {
						"type": "string"
					},
					"shifters": {
						"type": "string"
					},
					"speed": {
						"type": "string"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					},
					"updated": {
						"type": "string",
						"format": "date-time"
					},
					"createdBy": {
						"type": "string"
					},
					"updatedBy": {
						"type": "string"
					},
					"product": {
						"$ref": "#/components/schemas/Product"
					}
				},
				"required": [
					"id",
					"gear",
					"wheelSize",
					"category",
					"frameSize",
					"productUser",
					"shifters",
					"speed"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Vendor": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"businessName": {
						"type": "string"
					},
					"ownerFirstName": {
						"type": "string"
					},
					"ownerLastName": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"pancard": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"pincode": {
						"type": "string"
					},
					"gstNumber": {
						"type": "string"
					},
					"dateOfEstablishment": {
						"type": "string",
						"format": "date-time"
					},
					"contact": {
						"type": "string"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					},
					"updated": {
						"type": "string",
						"format": "date-time"
					},
					"createdBy": {
						"type": "string"
					},
					"updatedBy": {
						"type": "string"
					},
					"active": {
						"type": "boolean"
					},
					"vouchers": {
						"items": {
							"$ref": "#/components/schemas/Product"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"businessName",
					"contact",
					"active"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PurchaseLineItem": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"purchaseHeader": {
						"$ref": "#/components/schemas/PurchaseHeader"
					},
					"inventoryItem": {
						"items": {
							"$ref": "#/components/schemas/InventoryItem"
						},
						"type": "array"
					},
					"product": {
						"$ref": "#/components/schemas/Product"
					},
					"description": {
						"type": "string"
					},
					"qty": {
						"type": "number",
						"format": "double"
					},
					"rate": {
						"type": "number",
						"format": "double"
					},
					"amount": {
						"type": "number",
						"format": "double"
					},
					"cGstValue": {
						"type": "number",
						"format": "double"
					},
					"sGstValue": {
						"type": "number",
						"format": "double"
					},
					"cGstPercentage": {
						"type": "number",
						"format": "double"
					},
					"sGstPercentage": {
						"type": "number",
						"format": "double"
					},
					"packingCost": {
						"type": "number",
						"format": "double"
					},
					"distributorCost": {
						"type": "number",
						"format": "double"
					},
					"transportCost": {
						"type": "number",
						"format": "double"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					},
					"updated": {
						"type": "string",
						"format": "date-time"
					},
					"createdBy": {
						"type": "string"
					},
					"updatedBy": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"purchaseHeader",
					"inventoryItem",
					"description",
					"qty",
					"rate",
					"amount"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PurchaseHeader": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"invoiceNumber": {
						"type": "string"
					},
					"invoiceDate": {
						"type": "string",
						"format": "date-time"
					},
					"billTotal": {
						"type": "number",
						"format": "double"
					},
					"discountTotal": {
						"type": "number",
						"format": "double"
					},
					"packingTotal": {
						"type": "number",
						"format": "double"
					},
					"distributorTotal": {
						"type": "number",
						"format": "double"
					},
					"transportTotal": {
						"type": "number",
						"format": "double"
					},
					"cGstTotal": {
						"type": "number",
						"format": "double"
					},
					"sGstTotal": {
						"type": "number",
						"format": "double"
					},
					"roundOff": {
						"type": "number",
						"format": "double"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					},
					"updated": {
						"type": "string",
						"format": "date-time"
					},
					"createdBy": {
						"type": "string"
					},
					"updatedBy": {
						"type": "string"
					},
					"vendor": {
						"$ref": "#/components/schemas/Vendor"
					},
					"lineItems": {
						"items": {
							"$ref": "#/components/schemas/PurchaseLineItem"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"invoiceNumber",
					"invoiceDate",
					"vendor"
				],
				"type": "object",
				"additionalProperties": false
			},
			"InventoryItem": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"purchaseLineItem": {
						"$ref": "#/components/schemas/PurchaseLineItem"
					},
					"product": {
						"$ref": "#/components/schemas/Product"
					},
					"skuCode": {
						"type": "string"
					},
					"qty": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"type": "string"
					},
					"mrp": {
						"type": "number",
						"format": "double"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					},
					"updated": {
						"type": "string",
						"format": "date-time"
					},
					"createdBy": {
						"type": "string"
					},
					"updatedBy": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"purchaseLineItem",
					"skuCode",
					"qty",
					"mrp"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IBrandPaylod": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"url": {
						"type": "string"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					},
					"updated": {
						"type": "string",
						"format": "date-time"
					},
					"createdBy": {
						"type": "string"
					},
					"updatedBy": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"code",
					"url"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IProductType": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"productTypeCategory": {
						"$ref": "#/components/schemas/ProductTypeCategory"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					},
					"updated": {
						"type": "string",
						"format": "date-time"
					},
					"createdBy": {
						"type": "string"
					},
					"updatedBy": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"name",
					"productTypeCategory"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IProductPaylod": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"year": {
						"type": "number",
						"format": "double"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					},
					"updated": {
						"type": "string",
						"format": "date-time"
					},
					"createdBy": {
						"type": "string"
					},
					"updatedBy": {
						"type": "string"
					},
					"productDetail": {
						"$ref": "#/components/schemas/ProductDetail"
					},
					"productType": {
						"$ref": "#/components/schemas/ProductType"
					},
					"brand": {
						"$ref": "#/components/schemas/Brand"
					}
				},
				"required": [
					"id",
					"name",
					"year",
					"productType",
					"brand"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IVendorPaylod": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"businessName": {
						"type": "string"
					},
					"ownerFirstName": {
						"type": "string"
					},
					"ownerLastName": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"pancard": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"pincode": {
						"type": "string"
					},
					"gstNumber": {
						"type": "string"
					},
					"dateOfEstablishment": {
						"type": "string",
						"format": "date-time"
					},
					"contact": {
						"type": "string"
					},
					"active": {
						"type": "boolean"
					},
					"created": {
						"type": "string",
						"format": "date-time"
					},
					"updated": {
						"type": "string",
						"format": "date-time"
					},
					"createdBy": {
						"type": "string"
					},
					"updatedBy": {
						"type": "string"
					}
				},
				"required": [
					"id",
					"businessName",
					"ownerFirstName",
					"ownerLastName",
					"address",
					"pancard",
					"city",
					"pincode",
					"gstNumber",
					"dateOfEstablishment",
					"contact",
					"active"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "access_token",
				"in": "query"
			},
			"tsoa_auth": {
				"type": "oauth2",
				"flows": {
					"implicit": {
						"authorizationUrl": "http://swagger.io/api/oauth/dialog",
						"scopes": {}
					}
				}
			}
		}
	},
	"info": {
		"title": "palghar-milers",
		"version": "1.1.1",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/brands": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Brand"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Brand"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateData",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Brand"
								}
							}
						}
					}
				},
				"tags": [
					"Brand"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IBrandPaylod"
							}
						}
					}
				}
			}
		},
...
		"/product-types": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/ProductType"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"ProductType"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateData",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ProductType"
								}
							}
						}
					}
				},
				"tags": [
					"ProductType"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IProductType"
							}
						}
					}
				}
			}
		},
...
		"/product": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Product"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Product"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateData",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Product"
								}
							}
						}
					}
				},
				"tags": [
					"Product"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IProductPaylod"
							}
						}
					}
				}
			}
		},
		"/product/{id}": {
			"get": {
				"operationId": "GetOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Product"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Product"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/product/detail": {
			"post": {
				"operationId": "CreateWithDetail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Product"
								}
							}
						}
					}
				},
				"tags": [
					"Product"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IProductPaylod"
							}
						}
					}
				}
			}
		},
...
		"/vendors": {
			"get": {
				"operationId": "GetAll",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Vendor"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Vendors"
				],
				"security": [],
				"parameters": []
			},
			"post": {
				"operationId": "CreateData",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Vendor"
								}
							}
						}
					}
				},
				"tags": [
					"Vendors"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IVendorPaylod"
							}
						}
					}
				}
			}
		},
		"/vendors/{id}": {
			"get": {
				"operationId": "GetOne",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"allOf": [
										{
											"$ref": "#/components/schemas/Vendor"
										}
									],
									"nullable": true
								}
							}
						}
					}
				},
				"tags": [
					"Vendors"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}