# src/models/domain-enums.ts
export enum ProductTypeCategory {
    BIKES = "Bikes",
    RIDER_ACCESSORIES = "Rider Accessories",
    BIKE_ACCESSORIES = "Bike Accessories",
    COMPONENTS = "Components",
}

export enum Category {
    KIDS = "Kids",
    HYBRID = "Rider Accessories",
    MTB = "Bike Accessories",
    CITY = "City",
    ROAD = "Road",
    E_BIKE = "E-Bikes"
};

export enum SizeMaster {
    S = 'Small',
    M = 'Medium',
    L = 'Large',
    O = 'Other'
};
export enum BikeUsers {
    MEN = 'Men',
    WOMEN = 'Women',
    KIDS = 'Kids',
    UNISEX = 'Unisex'
};

export enum SuspensionType {
    RIGID = 'Rigid',
    SUSPENSION = 'Suspension',
    HYDRAULIC = 'Hydraulic'
};

export enum Brakes {
    DISC = 'Disc',
    POWER_V = 'Power-V Brakes',
};

export enum Material {
    ALLOY = 'Alloy',
    STEEL = 'Steel',
    CARBON = 'Carbon'
};

export enum ProductStatus {
    IN_SHIPMENT = 'In Shipment',
    FOR_SALE = 'For Sale',
    DEFECTIVE = 'Defective',
    UNWANTED = 'Unwanted',
    RETURNED = 'Returned',
    SOLD = 'sold',
};

// export const WheelSize = [14, 16, 18, 24, 26, 27, 27.5, 29];